migrate to long primary keys

+++ source/pony.core/src/main/java/net/dorokhov/pony/core/dao/AlbumDao.java
@@ -10,7 +10,7 @@ import java.util.List;
 /**
  * Album DAO.
  */
-public interface AlbumDao extends PagingAndSortingRepository<Album, Integer> {
+public interface AlbumDao extends PagingAndSortingRepository<Album, Long> {
 
 	/**
 	 * Retrieves count of albums by artist ID.
@@ -18,7 +18,7 @@ public interface AlbumDao extends PagingAndSortingRepository<Album, Integer> {
 	 * @param aArtistId artist ID
 	 * @return number of albums with the given artist ID
 	 */
-	public long countByArtistId(Integer aArtistId);
+	public long countByArtistId(Long aArtistId);
 
 	/**
 	 * Retrieves number of albums by artwork stored file ID.
@@ -26,7 +26,7 @@ public interface AlbumDao extends PagingAndSortingRepository<Album, Integer> {
 	 * @param aStoredFileId stored file ID
 	 * @return number of albums with the given artwork stored file ID
 	 */
-	public long countByArtworkId(Integer aStoredFileId);
+	public long countByArtworkId(Long aStoredFileId);
 
 	/**
 	 * Finds album by ID.
@@ -37,7 +37,7 @@ public interface AlbumDao extends PagingAndSortingRepository<Album, Integer> {
 	@Query("SELECT a FROM Album a " +
 			"INNER JOIN FETCH a.artist " +
 			"WHERE a.id = ?1")
-	public Album findById(Integer aId);
+	public Album findById(Long aId);
 
 	/**
 	 * Finds albums by artwork stored file ID.
@@ -46,7 +46,7 @@ public interface AlbumDao extends PagingAndSortingRepository<Album, Integer> {
 	 * @param aSort album sort option
 	 * @return list of albums with the given artwork stored file ID
 	 */
-	public List<Album> findByArtworkId(Integer aStoredFileId, Sort aSort);
+	public List<Album> findByArtworkId(Long aStoredFileId, Sort aSort);
 
 	/**
 	 * Finds albums by artist ID.
@@ -55,7 +55,7 @@ public interface AlbumDao extends PagingAndSortingRepository<Album, Integer> {
 	 * @param aSort album sort option
 	 * @return list of albums with the given artist ID
 	 */
-	public List<Album> findByArtistId(Integer aArtistId, Sort aSort);
+	public List<Album> findByArtistId(Long aArtistId, Sort aSort);
 
 	/**
 	 * Finds album by artist ID and album name.
@@ -64,6 +64,6 @@ public interface AlbumDao extends PagingAndSortingRepository<Album, Integer> {
 	 * @param aName album name
 	 * @return album with the given artist ID and album name or null if none found
 	 */
-	public Album findByArtistIdAndName(Integer aArtistId, String aName);
+	public Album findByArtistIdAndName(Long aArtistId, String aName);
 
 }
