+++ source/pony.core/src/main/java/net/dorokhov/pony/core/service/entity/SongFileServiceImpl.java
@@ -7,11 +7,11 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
-public class SongFileServiceImpl extends AbstractEntityService<SongFile, Integer, SongFileDao> implements SongFileService {
+public class SongFileServiceImpl extends AbstractEntityService<SongFile, Long, SongFileDao> implements SongFileService {
 
 	@Override
 	@Transactional(readOnly = true)
-	public long getCountByArtwork(Integer aStoredFileId) {
+	public long getCountByArtwork(Long aStoredFileId) {
 		return dao.countByArtworkId(aStoredFileId);
 	}
