+++ source/pony.core/src/main/resources/net/dorokhov/pony/core/dao/mysql/install.sql
@@ -1,6 +1,6 @@
 CREATE TABLE installation (
 
-	id INT NOT NULL AUTO_INCREMENT,
+	id BIGINT NOT NULL AUTO_INCREMENT,
 
 	creation_date TIMESTAMP NOT NULL,
 	update_date TIMESTAMP NOT NULL,
@@ -30,7 +30,7 @@ CREATE TABLE configuration (
 
 CREATE TABLE stored_file (
 
-	id INT NOT NULL AUTO_INCREMENT,
+	id BIGINT NOT NULL AUTO_INCREMENT,
 
 	creation_date TIMESTAMP NOT NULL,
 	update_date TIMESTAMP NOT NULL,
@@ -56,7 +56,7 @@ CREATE INDEX index_stored_file_tag_checksum ON stored_file(tag, checksum);
 
 CREATE TABLE song_file (
 
-	id INT NOT NULL AUTO_INCREMENT,
+	id BIGINT NOT NULL AUTO_INCREMENT,
 
 	creation_date TIMESTAMP NOT NULL,
 	update_date TIMESTAMP NOT NULL,
@@ -84,7 +84,7 @@ CREATE TABLE song_file (
 
 	year INT,
 
-	artwork_stored_file_id INT,
+	artwork_stored_file_id BIGINT,
 
 	FOREIGN KEY (artwork_stored_file_id) REFERENCES stored_file(id) ON DELETE SET NULL ON UPDATE CASCADE,
 
@@ -97,7 +97,7 @@ CREATE INDEX index_song_file_track_number_name ON song_file(disc_number, track_n
 
 CREATE TABLE artist (
 
-	id INT NOT NULL AUTO_INCREMENT,
+	id BIGINT NOT NULL AUTO_INCREMENT,
 
 	creation_date TIMESTAMP NOT NULL,
 	update_date TIMESTAMP NOT NULL,
@@ -105,7 +105,7 @@ CREATE TABLE artist (
 	generation BIGINT NOT NULL,
 
 	name VARCHAR(255) NOT NULL,
-	artwork_stored_file_id INT,
+	artwork_stored_file_id BIGINT,
 
 	FOREIGN KEY (artwork_stored_file_id) REFERENCES stored_file(id) ON DELETE SET NULL ON UPDATE CASCADE,
 
@@ -116,7 +116,7 @@ CREATE TABLE artist (
 
 CREATE TABLE album (
 
-	id INT NOT NULL AUTO_INCREMENT,
+	id BIGINT NOT NULL AUTO_INCREMENT,
 
 	creation_date TIMESTAMP NOT NULL,
 	update_date TIMESTAMP NOT NULL,
@@ -125,9 +125,9 @@ CREATE TABLE album (
 
 	name VARCHAR(255) NOT NULL,
 	year INT,
-	artwork_stored_file_id INT,
+	artwork_stored_file_id BIGINT,
 
-	artist_id INT NOT NULL,
+	artist_id BIGINT NOT NULL,
 
 	FOREIGN KEY (artist_id) REFERENCES artist(id) ON DELETE CASCADE ON UPDATE CASCADE,
 	FOREIGN KEY (artwork_stored_file_id) REFERENCES stored_file(id) ON DELETE SET NULL ON UPDATE CASCADE,
@@ -142,15 +142,15 @@ CREATE INDEX index_album_artist_id_year_name ON album(artist_id, year, name);
 
 CREATE TABLE song (
 
-	id INT NOT NULL AUTO_INCREMENT,
+	id BIGINT NOT NULL AUTO_INCREMENT,
 
 	creation_date TIMESTAMP NOT NULL,
 	update_date TIMESTAMP NOT NULL,
 
 	generation BIGINT NOT NULL,
 
-	song_file_id INT NOT NULL,
-	album_id INT NOT NULL,
+	song_file_id BIGINT NOT NULL,
+	album_id BIGINT NOT NULL,
 
 	UNIQUE (song_file_id),
 	PRIMARY KEY (id),
