+++ source/pony.web/src/main/java/net/dorokhov/pony/web/client/service/ArtistService.java
@@ -15,7 +15,7 @@ public interface ArtistService extends RemoteService {
 
 	public ArrayList<ArtistDto> search(String aQuery);
 
-	public ArtistDto getById(Integer aId);
+	public ArtistDto getById(Long aId);
 	public ArtistDto getByName(String aName);
 	public ArtistDto getByIdOrName(String aNameOrId);
