+++ source/pony.web/src/main/java/net/dorokhov/pony/web/client/service/ArtistServiceAsync.java
@@ -14,7 +14,7 @@ public interface ArtistServiceAsync {
 
 	public Request search(String aQuery, AsyncCallback<ArrayList<ArtistDto>> aCallback);
 
-	public Request getById(Integer aId, AsyncCallback<ArtistDto> aCallback);
+	public Request getById(Long aId, AsyncCallback<ArtistDto> aCallback);
 	public Request getByName(String aName, AsyncCallback<ArtistDto> aCallback);
 	public Request getByIdOrName(String aNameOrId, AsyncCallback<ArtistDto> aCallback);
