+++ source/pony.web/src/main/java/net/dorokhov/pony/web/client/service/SongService.java
@@ -9,14 +9,14 @@ import java.util.ArrayList;
 @RemoteServiceRelativePath("rpc/songService")
 public interface SongService extends RemoteService {
 
-	public Long getCountByAlbum(Integer aAlbumId);
-	public Long getCountByArtist(Integer aArtistId);
+	public Long getCountByAlbum(Long aAlbumId);
+	public Long getCountByArtist(Long aArtistId);
 
-	public ArrayList<SongDto> getByAlbum(Integer aAlbumId);
-	public ArrayList<SongDto> getByArtist(Integer aArtistId);
+	public ArrayList<SongDto> getByAlbum(Long aAlbumId);
+	public ArrayList<SongDto> getByArtist(Long aArtistId);
 
 	public ArrayList<SongDto> search(String aQuery);
 
-	public SongDto getById(Integer aId);
+	public SongDto getById(Long aId);
 
 }
